[NO-JIRA] Fix relative imports broken by the packages/backpack-web move
After moving sources from packages/<pkg>/ (depth 3 from repo root) to
packages/backpack-web/src/<pkg>/ (depth 5), every relative import that
reaches *outside* the moved tree needs two extra ../ segments. Updates:

- 32 *.stories.{ts,tsx,js} files importing from
  ../../../.storybook/bpk-storybook-utils → ../../../../../.storybook/...
- packages/backpack-web/src/bpk-stylesheets/webpack.config.babel.js
  ../../scripts/webpack/postCssPlugins → ../../../../scripts/...

Intra-tree imports (../../../bpk-component-X) still resolve correctly
because both source and target moved together.
