+++ packages/backpack-web/src/bpk-component-badge/src/BpkBadge.stories.tsx
@@ -34,7 +34,7 @@ import {
 } from '@skyscanner/bpk-foundations-web/tokens/base.es6';
 
 // @ts-expect-error Untyped import. See `decisions/imports-ts-suppressions.md`.
-import { BpkDarkExampleWrapper } from '../../../.storybook/bpk-storybook-utils';
+import { BpkDarkExampleWrapper } from '../../../../../.storybook/bpk-storybook-utils';
 import BpkSmallExclamationIcon from '../../bpk-component-icon/sm/exclamation';
 import BpkSmallHelpCircleIcon from '../../bpk-component-icon/sm/help-circle';
 import BpkSmallTickIcon from '../../bpk-component-icon/sm/tick-circle';
