+++ packages/backpack-web/src/bpk-component-banner-alert/src/BpkBannerAlert.stories.js
@@ -27,7 +27,7 @@ import BpkBannerAlert, {
   BpkBannerAlertExpandable,
   withBannerAlertState,
 } from '..';
-import { action } from '../../../.storybook/bpk-storybook-utils';
+import { action } from '../../../../../.storybook/bpk-storybook-utils';
 import AriaLiveDemo from '../../bpk-component-aria-live/src/BpkAriaLive.story-helpers';
 import BpkButton from '../../bpk-component-button';
 import CurrencyIcon from '../../bpk-component-icon/sm/currency';
