+++ packages/backpack-web/src/bpk-component-button/src/BpkButton.stories.tsx
@@ -18,7 +18,7 @@
 
 
 // @ts-expect-error Untyped import. See `decisions/imports-ts-suppressions.md`.
-import { action, BpkDarkExampleWrapper } from '../../../.storybook/bpk-storybook-utils';
+import { action, BpkDarkExampleWrapper } from '../../../../../.storybook/bpk-storybook-utils';
 import {
   withButtonAlignment,
   withLargeButtonAlignment,
