+++ packages/backpack-web/src/bpk-component-calendar/src/BpkCalendar.stories-components.js
@@ -39,7 +39,7 @@ import BpkCalendar, {
   CustomPropTypes,
   CALENDAR_SELECTION_TYPE,
 } from '..';
-import { action } from '../../../.storybook/bpk-storybook-utils';
+import { action } from '../../../../../.storybook/bpk-storybook-utils';
 import BpkButton from '../../bpk-component-button';
 import {
   withButtonAlignment,
