+++ packages/backpack-web/src/bpk-component-close-button/src/BpkCloseButton.stories.tsx
@@ -17,7 +17,7 @@
  */
 
 // @ts-expect-error Untyped import. See `decisions/imports-ts-suppressions.md`.
-import { action } from '../../../.storybook/bpk-storybook-utils';
+import { action } from '../../../../../.storybook/bpk-storybook-utils';
 
 import BpkCloseButton from './BpkCloseButton';
