+++ packages/backpack-web/src/bpk-component-datepicker/src/BpkDatepicker.stories.js
@@ -28,7 +28,7 @@ import {
   colorSkyGray,
 } from '@skyscanner/bpk-foundations-web/tokens/base.es6';
 
-import { action } from '../../../.storybook/bpk-storybook-utils';
+import { action } from '../../../../../.storybook/bpk-storybook-utils';
 import {
   BpkCalendarNav,
   BpkCalendarGridHeader,
