+++ packages/backpack-web/src/bpk-component-horizontal-nav/src/BpkHorizontalNav.stories.js
@@ -23,7 +23,7 @@ import {
   colorSkyGrayTint06,
 } from '@skyscanner/bpk-foundations-web/tokens/base.es6';
 
-import { BpkDarkExampleWrapper } from '../../../.storybook/bpk-storybook-utils';
+import { BpkDarkExampleWrapper } from '../../../../../.storybook/bpk-storybook-utils';
 import BpkText from '../../bpk-component-text';
 import { cssModules } from '../../bpk-react-utils';
