+++ packages/backpack-web/src/bpk-component-navigation-bar/src/BpkNavigationBar.stories.js
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-import { action } from '../../../.storybook/bpk-storybook-utils';
+import { action } from '../../../../../.storybook/bpk-storybook-utils';
 import { withRtlSupport } from '../../bpk-component-icon';
 import CloseIcon from '../../bpk-component-icon/sm/close';
 import ArrowIcon from '../../bpk-component-icon/sm/long-arrow-left';
