+++ packages/backpack-web/src/bpk-component-radio/src/BpkRadio.stories.tsx
@@ -19,7 +19,7 @@
 import { Component } from 'react';
 
 // @ts-expect-error Untyped import. See `decisions/imports-ts-suppressions.md`.
-import { action, BpkDarkExampleWrapper } from '../../../.storybook/bpk-storybook-utils';
+import { action, BpkDarkExampleWrapper } from '../../../../../.storybook/bpk-storybook-utils';
 
 import BpkRadio from './BpkRadio';
