+++ packages/backpack-web/src/bpk-component-scrollable-calendar/src/BpkScrollableCalendar.stories.js
@@ -22,7 +22,7 @@ import { Component, useEffect } from 'react';
 import { endOfMonth } from 'date-fns/endOfMonth';
 import { startOfMonth } from 'date-fns/startOfMonth';
 
-import { action } from '../../../.storybook/bpk-storybook-utils';
+import { action } from '../../../../../.storybook/bpk-storybook-utils';
 import {
   DateUtils,
   CustomPropTypes,
