+++ packages/backpack-web/src/bpk-component-segmented-control/src/BpkSegmentedControl.stories.tsx
@@ -24,7 +24,7 @@ import {
 } from '@skyscanner/bpk-foundations-web/tokens/base.es6';
 
 // @ts-expect-error Untyped import. See `decisions/imports-ts-suppressions.md`.
-import { BpkDarkExampleWrapper } from '../../../.storybook/bpk-storybook-utils';
+import { BpkDarkExampleWrapper } from '../../../../../.storybook/bpk-storybook-utils';
 import BpkText from '../../bpk-component-text';
 import { cssModules } from '../../bpk-react-utils';
