+++ packages/backpack-web/src/bpk-component-split-input/src/BpkSplitInput.stories.tsx
@@ -17,7 +17,7 @@
  */
 
 // @ts-expect-error Untyped import. See `decisions/imports-ts-suppressions.md`.
-import { action } from '../../../.storybook/bpk-storybook-utils';
+import { action } from '../../../../../.storybook/bpk-storybook-utils';
 import BpkLabel from '../../bpk-component-label';
 
 import BpkSplitInput, { INPUT_TYPES } from './BpkSplitInput';
