+++ packages/backpack-web/src/bpk-component-star-rating/src/BpkStarRating.stories.tsx
@@ -18,7 +18,7 @@
 
 
 // @ts-expect-error Untyped import
-import { action } from '../../../.storybook/bpk-storybook-utils';
+import { action } from '../../../../../.storybook/bpk-storybook-utils';
 import {
   BpkTable,
   BpkTableBody,
