+++ packages/backpack-web/src/bpk-component-swap-button/src/BpkSwapButton.stories.tsx
@@ -17,7 +17,7 @@
  */
 
 // @ts-expect-error Untyped import. See `decisions/imports-ts-suppressions.md`.
-import { action, BpkDarkExampleWrapper } from '../../../.storybook/bpk-storybook-utils';
+import { action, BpkDarkExampleWrapper } from '../../../../../.storybook/bpk-storybook-utils';
 
 import BpkSwapButton, { SWAPBUTTON_STYLES } from './BpkSwapButton';
