+++ packages/backpack-web/src/bpk-component-textarea/src/BpkTextarea.stories.tsx
@@ -17,7 +17,7 @@
  */
 
 // @ts-expect-error Untyped import. See `decisions/imports-ts-suppressions.md`.
-import { action } from '../../../.storybook/bpk-storybook-utils';
+import { action } from '../../../../../.storybook/bpk-storybook-utils';
 import BpkLabel from '../../bpk-component-label';
 
 import BpkTextarea from './BpkTextarea';
